refactor(features): move shared classes to common dir
