+++ app/src/main/java/com/leehendryp/codechallenge/core/data/database/CodeChallengeDatabase.kt
@@ -2,8 +2,8 @@ package com.leehendryp.codechallenge.core.data.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.leehendryp.codechallenge.features.list.data.local.AlbumDao
-import com.leehendryp.codechallenge.features.list.data.local.model.AlbumEntity
+import com.leehendryp.codechallenge.features.common.data.local.AlbumDao
+import com.leehendryp.codechallenge.features.common.data.local.model.AlbumEntity
 
 @Database(entities = [AlbumEntity::class], version = 1)
 abstract class CodeChallengeDatabase : RoomDatabase() {
