+++ app/src/main/java/com/leehendryp/codechallenge/features/common/data/AlbumRemoteMediator.kt
@@ -1,12 +1,12 @@
-package com.leehendryp.codechallenge.features.list.data
+package com.leehendryp.codechallenge.features.common.data
 
 import androidx.paging.ExperimentalPagingApi
 import androidx.paging.LoadType
 import androidx.paging.PagingState
 import androidx.paging.RemoteMediator
-import com.leehendryp.codechallenge.features.list.data.local.LocalDataSource
-import com.leehendryp.codechallenge.features.list.data.local.model.AlbumEntity
-import com.leehendryp.codechallenge.features.list.data.remote.RemoteDataSource
+import com.leehendryp.codechallenge.features.common.data.local.LocalDataSource
+import com.leehendryp.codechallenge.features.common.data.local.model.AlbumEntity
+import com.leehendryp.codechallenge.features.common.data.remote.RemoteDataSource
 import kotlinx.coroutines.flow.firstOrNull
 
 @OptIn(ExperimentalPagingApi::class)
