+++ app/src/main/java/com/leehendryp/codechallenge/features/common/data/local/AlbumDao.kt
@@ -1,11 +1,11 @@
-package com.leehendryp.codechallenge.features.list.data.local
+package com.leehendryp.codechallenge.features.common.data.local
 
 import androidx.paging.PagingSource
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.leehendryp.codechallenge.features.list.data.local.model.AlbumEntity
+import com.leehendryp.codechallenge.features.common.data.local.model.AlbumEntity
 
 @Dao
 interface AlbumDao {
@@ -15,6 +15,6 @@ interface AlbumDao {
     @Query("SELECT * FROM album WHERE id = :id")
     fun getAlbum(id: Int): AlbumEntity
 
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    @Insert(onConflict = OnConflictStrategy.Companion.REPLACE)
     suspend fun insertAll(albums: List<AlbumEntity>)
 }
