+++ app/src/main/java/com/leehendryp/codechallenge/features/common/data/remote/RemoteDataSource.kt
@@ -1,0 +1,8 @@
+package com.leehendryp.codechallenge.features.common.data.remote
+
+import com.leehendryp.codechallenge.features.common.domain.Album
+import kotlinx.coroutines.flow.Flow
+
+internal interface RemoteDataSource {
+    fun fetchAlbums(): Flow<List<Album>>
+}
