+++ app/src/main/java/com/leehendryp/codechallenge/features/common/data/remote/model/AlbumResponse.kt
@@ -1,4 +1,4 @@
-package com.leehendryp.codechallenge.features.list.data.remote.model
+package com.leehendryp.codechallenge.features.common.data.remote.model
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
