--- app/src/main/java/com/leehendryp/codechallenge/features/list/data/remote/RemoteDataSource.kt
@@ -1,8 +1,0 @@
-package com.leehendryp.codechallenge.features.list.data.remote
-
-import com.leehendryp.codechallenge.features.list.domain.Album
-import kotlinx.coroutines.flow.Flow
-
-internal interface RemoteDataSource {
-    fun fetchAlbums(): Flow<List<Album>>
-}
