+++ app/src/test/java/com/leehendryp/codechallenge/features/details/presentation/ui/AlbumDetailsItemContentTest.kt
@@ -4,7 +4,7 @@ import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.onNodeWithText
-import com.leehendryp.codechallenge.features.list.domain.MockDomainModels
+import com.leehendryp.codechallenge.features.common.domain.MockDomainModels
 import com.leehendryp.codechallenge.ui.ds.DS_ASYNC_IMAGE_TEST_TAG
 import com.leehendryp.codechallenge.ui.theme.CodeChallengeTheme
 import org.junit.Rule
