+++ app/src/test/java/com/leehendryp/codechallenge/features/list/data/local/model/LocalDataMappersTest.kt
@@ -1,8 +1,11 @@
 package com.leehendryp.codechallenge.features.list.data.local.model
 
+import com.leehendryp.codechallenge.features.common.data.local.model.AlbumEntity
+import com.leehendryp.codechallenge.features.common.data.local.model.toDomainModel
+import com.leehendryp.codechallenge.features.common.data.local.model.toLocalDataModels
+import com.leehendryp.codechallenge.features.common.domain.Album
+import com.leehendryp.codechallenge.features.common.domain.MockDomainModels
 import com.leehendryp.codechallenge.features.list.data.model.MockDataModels
-import com.leehendryp.codechallenge.features.list.domain.Album
-import com.leehendryp.codechallenge.features.list.domain.MockDomainModels
 import org.hamcrest.CoreMatchers.equalTo
 import org.hamcrest.MatcherAssert
 import org.junit.Test
