+++ app/src/test/java/com/leehendryp/codechallenge/features/list/data/model/ToDomainMapperTest.kt
@@ -1,9 +1,9 @@
 package com.leehendryp.codechallenge.features.list.data.model
 
-import com.leehendryp.codechallenge.features.list.data.remote.model.AlbumResponse
-import com.leehendryp.codechallenge.features.list.data.remote.model.toDomainModels
-import com.leehendryp.codechallenge.features.list.domain.Album
-import com.leehendryp.codechallenge.features.list.domain.MockDomainModels
+import com.leehendryp.codechallenge.features.common.data.remote.model.AlbumResponse
+import com.leehendryp.codechallenge.features.common.data.remote.model.toDomainModels
+import com.leehendryp.codechallenge.features.common.domain.Album
+import com.leehendryp.codechallenge.features.common.domain.MockDomainModels
 import org.hamcrest.CoreMatchers.equalTo
 import org.hamcrest.MatcherAssert.assertThat
 import org.junit.Test
